gcp/requirements.txt
functions-framework==3.*
requests==2.*
sqlalchemy==2.*
cloud-sql-python-connector[pg8000]==1.*
gunicorn>=22.0.0
google-cloud-secret-manager==2.20.*
firebase-admin==6.5.*
marshmallow==3.21.*